+++ CHANGES.md
@@ -7,7 +7,10 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
 
 ## next (not yet released)
 
-No changes yet.
+- `login()` and `LoginError` now understand
+  that the `reason` returned by the API in case of failed login
+  can be an object rather than a string
+  (if the session’s `defaultParams` specify a non-default `errorformat`).
 
 ## v0.2.0 (2023-07-10)